+++ src/Infrastructure/Adapter/SessionAdapter.php
@@ -57,7 +57,7 @@ class SessionAdapter implements SessionInterface
     /**
      * {@inheritdoc}
      */
-    public function set($key, $data)
+    public function set($key, $data): void
     {
         $this->session->set($key, $data);
     }