+++ src/Ui/Bundle/AdminBundle/Form/Type/Participant/ImportType.php
@@ -1,13 +1,5 @@
 <?php
 
-/*
- * This file is part of the Proximum Vimeet project.
- *
- * Copyright (C) Proximum
- *
- * @author Elao <contact@elao.com>
- */
-
 namespace Proximum\Vimeet\Ui\Bundle\AdminBundle\Form\Type\Participant;
 
 use Proximum\Vimeet\Application\Command\Participant\Import;
@@ -16,6 +8,8 @@ use Proximum\Vimeet\Domain\Model\Event;
 use Proximum\Vimeet\Ui\Bundle\AdminBundle\Form\Type\Nomenclature\CharsetChoiceType;
 use Proximum\Vimeet\Ui\Bundle\AdminBundle\Form\Type\TypeChoiceType;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
+use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\Extension\Core\Type\FileType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
@@ -25,25 +19,35 @@ class ImportType extends AbstractType
     /**
      * {@inheritdoc}
      */
-    public function buildForm(FormBuilderInterface $builder, array $options)
+    public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
             ->add('type', TypeChoiceType::class, [
-                'event'    => $options['event'],
-                'locale'   => $options['locale'],
-                'user'     => $options['user'],
+                'event' => $options['event'],
+                'locale' => $options['locale'],
+                'user' => $options['user'],
                 'required' => true,
+                'placeholder' => 'form.participant_import.children.type.placeholder',
             ])
             ->add('file', FileType::class, [
                 'required' => true,
             ])
-            ->add('charset', CharsetChoiceType::class);
+            ->add('charset', CharsetChoiceType::class)
+            ->add('allowMultiSheet', ChoiceType::class, [
+                'expanded' => true,
+                'choices' => [
+                    'form.participant_import.children.allowMultiSheet.choice.yes' => true,
+                    'form.participant_import.children.allowMultiSheet.choice.no' => false
+                ],
+                'required' => true,
+            ])
+        ;
     }
 
     /**
      * {@inheritdoc}
      */
-    public function configureOptions(OptionsResolver $resolver)
+    public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setRequired(['event', 'locale', 'user']);
         $resolver->setAllowedTypes('event', Event::class);
@@ -57,7 +61,7 @@ class ImportType extends AbstractType
     /**
      * {@inheritdoc}
      */
-    public function getBlockPrefix()
+    public function getBlockPrefix(): string
     {
         return 'participant_import';
     }